backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
torch==2.8.0
torchvision==0.23.0
numpy==1.26.0
Pillow==10.0.1
python-jose==3.3.0
passlib==1.7.4
dlib==19.24.2
face-recognition==1.3.0
scikit-learn==1.3.2
